Agregado Endpoints de resenas

+++ db/migrate/20220121233150_create_clientes.rb
@@ -1,10 +1,11 @@
 class CreateClientes < ActiveRecord::Migration[7.0]
   def change
     create_table :clientes do |t|
-      t.decimal :historial
+      t.integer :historial
       t.references :usuario, null: false, foreign_key: true
 
       t.timestamps
+
     end
   end
 end
